+++ microhapulator/type.py
@@ -39,7 +39,7 @@ def check_index(bamfile):
         pysam.index(bamfile)
 
 
-def tally_haplotypes(bamfile, refrfasta):
+def tally_haplotypes(bamfile, refrfasta, minbasequal=10):
     totaldiscarded = 0
     with microhapulator.open(refrfasta, 'r') as fh:
         offsets = parse_variant_offsets_from_fasta_headers(fh)
@@ -51,7 +51,7 @@ def tally_haplotypes(bamfile, refrfasta):
         ht = defaultdict(dict)
         varloc = set(offsets[locusid])
         cov_pos = list()
-        for column in bam.pileup(locusid):
+        for column in bam.pileup(locusid, min_base_quality=minbasequal):
             cov_pos.append(column.n)
             if column.pos not in varloc:
                 continue
@@ -76,20 +76,20 @@ def tally_haplotypes(bamfile, refrfasta):
     )
 
 
-def type(bamfile, refrfasta, threshold=None):
-    genotyper = tally_haplotypes(bamfile, refrfasta)
+def type(bamfile, refrfasta, minbasequal=10, ecthreshold=0.25, static=None, dynamic=None):
+    genotyper = tally_haplotypes(bamfile, refrfasta, minbasequal=minbasequal)
     profile = microhapulator.profile.ObservedProfile()
     for locusid, cov_by_pos, htcounts, ndiscarded in genotyper:
         profile.record_coverage(locusid, cov_by_pos, ndiscarded=ndiscarded)
         for allele, count in htcounts.items():
             profile.record_allele(locusid, allele, count)
-        if threshold is None:
-            profile.data['markers'][locusid]['genotype'] = list()
-    if threshold is not None:
-        profile.infer(threshold=threshold)
+    profile.infer(ecthreshold=ecthreshold, static=static, dynamic=dynamic)
     return profile
 
 
 def main(args):
-    profile = type(args.bam, args.refr, threshold=args.threshold)
+    profile = type(
+        args.bam, args.refr, minbasequal=args.base_qual, ecthreshold=args.effcov,
+        static=args.static, dynamic=args.dynamic
+    )
     profile.dump(args.out)