+++ CHANGELOG.md
@@ -4,13 +4,17 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+### Added
+- New `--base-qual` parameter for `mhpl8r type` to set the minimum required base quality when iterating over reads in a pileup (#83).
+- New `mhpl8r balance` module for calculating and visualizing interlocus balance (#85).
+
 ### Changed
 - Updated mybinder demo (see #69).
 - Simulated Illumina sequencing now uses 1 thread by default, which paradoxically should lead to better performance (#71).
 - Moved panel definition code moved out of the core code and into dedicated notebooks (#74).
 - Replaced `MissingBAMIndexError` with BAM auto-indexing code (#78).
 - Improved read names and choice of interleaved or paired output for `mhpl8r seq` (#80).
-- Disable application of filtering thresholds by default in `mhpl8r type` (#82).
+- Replaced `--threshold` with `--static` and `--dynamic` in in `mhpl8r type`, disabled both by default (#82, #83).
 
 ### Fixed
 - Corrected a bug with Fastq headers in `mhpl8r seq` module (see #71).